Remove network dagger module
