+++ data/location/build.gradle
@@ -56,6 +56,7 @@ dependencies {
     implementation(libs.playServicesLocation)
 
     implementation(libs.okHttp)
+    implementation(libs.okHttpLogger)
     implementation(libs.retrofit)
     implementation(libs.retrofitGsonConverter)
