+++ data/location/src/main/java/akio/apps/myrun/data/location/LocationDataModule.kt
@@ -10,7 +10,7 @@ import akio.apps.myrun.data.location.impl.LocationDataSourceImpl
 import akio.apps.myrun.data.location.impl.PolyUtilImpl
 import akio.apps.myrun.data.location.impl.SphericalUtilImpl
 import akio.apps.myrun.data.location.impl.model.GoogleMapDirectionApiKey
-import akio.apps.myrun.wiring.common.NetworkModule
+import akio.apps.myrun.wiring.common.BuildConfig
 import android.app.Application
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.LocationServices
@@ -21,6 +21,7 @@ import dagger.Binds
 import dagger.Module
 import dagger.Provides
 import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 
@@ -28,7 +29,6 @@ import retrofit2.converter.gson.GsonConverterFactory
     includes = [
         LocationDataModule.Providers::class,
         DirectionDataModule::class,
-        NetworkModule::class
     ]
 )
 interface LocationDataModule {
@@ -54,7 +54,14 @@ interface LocationDataModule {
             LocationServices.getFusedLocationProviderClient(application)
 
         @Provides
-        fun googleDirectionApi(okHttpClientBuilder: OkHttpClient.Builder): GoogleMapDirectionApi {
+        fun googleDirectionApi(): GoogleMapDirectionApi {
+            val okHttpClientBuilder = OkHttpClient.Builder()
+            if (BuildConfig.DEBUG) {
+                val logger = HttpLoggingInterceptor().apply {
+                    level = HttpLoggingInterceptor.Level.BODY
+                }
+                okHttpClientBuilder.addInterceptor(logger)
+            }
             val okHttpClient = okHttpClientBuilder.build()
 
             val retrofit = Retrofit.Builder()
