+++ data/user/src/main/java/akio/apps/myrun/data/user/impl/FirebaseUserRecentPlaceRepository.kt
@@ -4,6 +4,7 @@ import akio.apps.myrun.data.common.Resource
 import akio.apps.myrun.data.user.api.PlaceIdentifier
 import akio.apps.myrun.data.user.api.UserRecentPlaceRepository
 import akio.apps.myrun.data.user.impl.model.FirestorePlaceIdentifier
+import akio.apps.myrun.wiring.common.NamedIoDispatcher
 import com.google.firebase.firestore.CollectionReference
 import com.google.firebase.firestore.FirebaseFirestore
 import javax.inject.Inject
@@ -21,7 +22,8 @@ import kotlinx.coroutines.withContext
 
 class FirebaseUserRecentPlaceRepository @Inject constructor(
     private val firestore: FirebaseFirestore,
-    @akio.apps.myrun.wiring.common.NamedIoDispatcher private val ioDispatcher: CoroutineDispatcher,
+    @NamedIoDispatcher
+    private val ioDispatcher: CoroutineDispatcher,
 ) : UserRecentPlaceRepository {
 
     private val recentPlaceCollection: CollectionReference
