+++ domain/src/main/java/akio/apps/myrun/domain/activity/ExportTempTcxFileUsecase.kt
@@ -17,7 +17,8 @@ class ExportTempTcxFileUsecase @Inject constructor(
     private val application: Application,
     private val activityTcxFileWriter: ActivityTcxFileWriter,
     private val activityRepository: ActivityRepository,
-    @NamedIoDispatcher private val ioDispatcher: CoroutineDispatcher,
+    @NamedIoDispatcher
+    private val ioDispatcher: CoroutineDispatcher,
 ) {
 
     private val timeFormatter = SimpleDateFormat("ddMMyyyy_HHmm", Locale.US)
